Mild Touch-Ups
Removing Commented out code
Adding comments and notes
Shifting order of execution in code, etc... Little stuff, to add quality of life and clarity to code
